Adding support for inline chat
This adds in support for loading as inline chat, with the default set to false to not affect existing users.

Tests added to confirm the setting is passed through if it's set to true, but not if it's set to false.

This will require a new setting to be toggled by the account owner at segment.com for setting the chat to inline chat mode.